feat: split create data assets into two new functions
