+++ src/BIAnalyticsService/CreateParticipantScreeningEpisode/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 {
   "profiles": {
-    "RetrieveData": {
+    "CreateParticipantScreeningEpisode": {
       "commandName": "Project",
       "commandLineArgs": "--port 7254",
       "launchBrowser": false
