+++ src/BIAnalyticsService/CreateParticipantScreeningEpisode/local.settings.json.template
@@ -1,0 +1,12 @@
+{
+  "IsEncrypted": false,
+  "Values": {
+    "AzureWebJobsStorage": "UseDevelopmentStorage=true",
+    "FUNCTIONS_WORKER_RUNTIME": "dotnet-isolated",
+    "GetEpisodeUrl": "http://localhost:6060/api/GetEpisode",
+    "CreateParticipantScreeningEpisodeUrl":"http://localhost:6010/api/CreateParticipantScreeningEpisode",
+  },
+  "Host": {
+    "LocalHttpPort": 6008
+  }
+}
