+++ src/BIAnalyticsService/CreateParticipantScreeningProfile/Program.cs
@@ -1,0 +1,12 @@
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using NHS.ServiceInsights.Common;
+
+var host = new HostBuilder()
+    .ConfigureFunctionsWorkerDefaults()
+    .ConfigureServices(services =>
+    {
+        services.AddSingleton<IHttpRequestService, HttpRequestService>();
+    })
+    .Build();
+host.Run();
