+++ src/BIAnalyticsService/CreateParticipantScreeningProfile/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+    "CreateParticipantScreeningProfile": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7254",
+      "launchBrowser": false
+    }
+  }
+}
